all the routes working

+++ routes/api/product-routes.js
@@ -1,65 +1,57 @@
-const router = require('express').Router();
-const { Product, Category, Tag, ProductTag } = require('../../models');
+const router = require("express").Router()
+const { Product, Category, Tag, ProductTag } = require("../../models")
 
 // The `/api/products` endpoint
 
 // get all products
-router.get('/', async (req, res) => {
-  // find all products
-  // be sure to include its associated Category and Tag data
-
+// find all products and associated Category and Tag data
+router.get("/", async (req, res) => {
   const productData = await Product.findAll({
-    include:[ 
+    include: [
       {
-      model: Category,
-      as: 'category'
-    },
-    {
-      model: Tag,
-      as: 'tags'
-    },
-  ],
-   
-  });
-  console.log('product get')
+        model: Category,
+        as: "category",
+      },
+      {
+        model: Tag,
+        as: "tags",
+      },
+    ],
+  })
+  console.log("product get")
   return res.json(productData)
 })
 
-
 // get one product
-router.get('/:id', async(req, res) => {
-  // find a single product by its `id`
-  // be sure to include its associated Category and Tag data
+router.get("/:id", async (req, res) => {
   try {
-    const productIdData = await Product.findByPk(req.params.id,
-       {
-        include:[ {
+    const productIdData = await Product.findByPk(req.params.id, {
+      include: [
+        {
           model: Category,
-          as: 'category'
+          as: "category",
         },
         {
           model: Tag,
-          as: 'tags'
-        }
-
-      ]
-    });
+          as: "tags",
+        },
+      ],
+    })
 
     if (!productIdData) {
-      res.status(404).json({ message: 'No product found with that id!' });
-      return;
+      res.status(404).json({ message: "No product found with that id!" })
+      return
     } else {
-    res.status(200).json(productIdData);
+      res.status(200).json(productIdData)
     }
   } catch (err) {
     console.log(err)
-    res.status(500).json(err);
+    res.status(500).json(err)
   }
-});
+})
 
 // create new product
-router.post('/', async (req, res) => {
-  /* req.body should look like this...
+/* req.body should look like this...
     // {
     //   product_name: "Basketball",
     //   price: 200.00,
@@ -67,6 +59,7 @@ router.post('/', async (req, res) => {
     //   tagIds: [1, 2, 3, 4]
     // }
   */
+router.post("/", async (req, res) => {
   Product.create(req.body)
     .then((product) => {
       // if there's product tags, we need to create pairings to bulk create in the ProductTag model
@@ -75,23 +68,22 @@ router.post('/', async (req, res) => {
           return {
             product_id: product.id,
             tag_id,
-          };
-        });
-        return ProductTag.bulkCreate(productTagIdArr);
+          }
+        })
+        return ProductTag.bulkCreate(productTagIdArr)
       }
       // if no product tags, just respond
-      res.status(200).json(product);
+      res.status(200).json(product)
     })
     .then((productTagIds) => res.status(200).json(productTagIds))
     .catch((err) => {
-      console.log(err);
-      res.status(400).json(err);
-    });
-});
+      console.log(err)
+      res.status(400).json(err)
+    })
+})
 
 // update product
-router.put('/:id', (req, res) => {
-  // update product data
+router.put("/:id", (req, res) => {
   Product.update(req.body, {
     where: {
       id: req.params.id,
@@ -99,40 +91,57 @@ router.put('/:id', (req, res) => {
   })
     .then((product) => {
       // find all associated tags from ProductTag
-      return ProductTag.findAll({ where: { product_id: req.params.id } });
+      return ProductTag.findAll({ where: { product_id: req.params.id } })
     })
     .then((productTags) => {
       // get list of current tag_ids
-      const productTagIds = productTags.map(({ tag_id }) => tag_id);
-      // create filtered list of new tag_ids
-      const newProductTags = req.body.tagIds
-        .filter((tag_id) => !productTagIds.includes(tag_id))
-        .map((tag_id) => {
-          return {
-            product_id: req.params.id,
-            tag_id,
-          };
-        });
-      // figure out which ones to remove
-      const productTagsToRemove = productTags
-        .filter(({ tag_id }) => !req.body.tagIds.includes(tag_id))
-        .map(({ id }) => id);
+      if (req.body.tagIds) {
+        const productTagIds = productTags.map(({ tag_id }) => tag_id)
+        // create filtered list of new tag_ids
+        const newProductTags = req.body.tagIds
+          .filter((tag_id) => !productTagIds.includes(tag_id))
+          .map((tag_id) => {
+            return {
+              product_id: req.params.id,
+              tag_id,
+            }
+          })
+        // figure out which ones to remove
+        const productTagsToRemove = productTags
+          .filter(({ tag_id }) => !req.body.tagIds.includes(tag_id))
+          .map(({ id }) => id)
 
-      // run both actions
-      return Promise.all([
-        ProductTag.destroy({ where: { id: productTagsToRemove } }),
-        ProductTag.bulkCreate(newProductTags),
-      ]);
+        // run both actions
+        return Promise.all([
+          ProductTag.destroy({ where: { id: productTagsToRemove } }),
+          ProductTag.bulkCreate(newProductTags),
+        ])
+      }
+      return []
     })
     .then((updatedProductTags) => res.json(updatedProductTags))
     .catch((err) => {
-      // console.log(err);
-      res.status(400).json(err);
-    });
-});
+      console.log(err)
+      res.status(400).json(err)
+    })
+})
 
-router.delete('/:id', (req, res) => {
-  // delete one product by its `id` value
-});
+// delete one product by its `id` value
+router.delete("/:id", async (req, res) => {
+  try {
+    const userData = await Product.destroy({
+      where: {
+        id: req.params.id,
+      },
+    })
+    if (!userData) {
+      res.status(404).json({ message: "No product with this id!" })
+      return
+    }
+    res.status(200).json(userData)
+  } catch (err) {
+    res.status(500).json(err)
+  }
+})
 
-module.exports = router;
+module.exports = router
